requirements.txt
elasticsearch>=7.5.1,<8
python-dotenv>=0.10.5,<0.11
certifi
elasticsearch-dsl>=7,<8